OGUZCAN

+++ resources/views/contact/show.blade.php
@@ -5,13 +5,6 @@
 @section('content')
 
     <div class="row">
-        <div class="col-md-12">
-            <h2>
-                <a href="{{ route('contact.index') }}">Go back</a>
-            </h2>
-        </div>
-    </div>
-    <div class="row">
         <h3>
             {{ $contact->firstname }} {{ $contact->insertion }} {{ $contact->surename }}
         </h3>
@@ -38,10 +31,17 @@
             @endif
         </div>
     </div>
+
     <div>
-        <a href="{{ route('contact.edit', $contact->id) }}" class="btn btn-default"> <span class="glyphicon glyphicon-pencil"></span></a>
-        {!! Form::open(['route' => ['contact.destroy', $contact->id], 'method' => 'delete', 'style' => 'display:inline']) !!}
-        {!! Form::submit("trash", ['class' => 'btn btn-danger delete']) !!}
-        {!! Form::close() !!}
+        @if(Auth::check())
+                @if(Auth::user()->contact_id == $contact->id OR Auth::user()->getRole() == 'admin' )
+                    <a href="{{ route('contact.edit', $contact->id) }}" class="btn btn-default"> <span class="glyphicon glyphicon-pencil"></span></a>
+                @endif
+            @if(Auth::user()->getRole() == 'admin' )
+                {!! Form::open(['route' => ['contact.destroy', $contact->id], 'method' => 'delete', 'style' => 'display:inline']) !!}
+                {!! Form::submit("trash", ['class' => 'btn btn-danger delete']) !!}
+                {!! Form::close() !!}
+            @endif
+        @endif
     </div>
 @stop
